move calcsep to basic_util
